+++ src/gui/item_models/property_model_full.hpp
@@ -22,6 +22,7 @@ public:
         ColumnColor,
         ColumnVisible,
         ColumnLocked,
+        ColumnMask,
 
         ColumnCount,